Addendum for YARN-6102.
