+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/RMContextImpl.java
@@ -238,6 +238,17 @@ public class RMContextImpl implements RMContext {
   }
 
   @Override
+  public RMTimelineCollectorManager getRMTimelineCollectorManager() {
+    return serviceContext.getRMTimelineCollectorManager();
+  }
+
+  @Override
+  public void setRMTimelineCollectorManager(
+      RMTimelineCollectorManager timelineCollectorManager) {
+    serviceContext.setRMTimelineCollectorManager(timelineCollectorManager);
+  }
+
+  @Override
   public ConfigurationProvider getConfigurationProvider() {
     return serviceContext.getConfigurationProvider();
   }
@@ -457,18 +468,6 @@ public class RMContextImpl implements RMContext {
   @Override
   public boolean isWorkPreservingRecoveryEnabled() {
     return activeServiceContext.isWorkPreservingRecoveryEnabled();
-  }
-
-  @Override
-  public void setRMTimelineCollectorManager(
-      RMTimelineCollectorManager timelineCollectorManager) {
-    activeServiceContext.setRMTimelineCollectorManager(
-        timelineCollectorManager);
-  }
-
-  @Override
-  public RMTimelineCollectorManager getRMTimelineCollectorManager() {
-    return activeServiceContext.getRMTimelineCollectorManager();
   }
 
   @Override
